Выносит служелные функции
