+++ src/components/trip-event.js
@@ -1,12 +1,7 @@
-import {correctDateFormat} from '../mock/way-point.js';
+import {correctDateFormat} from '../utils.js';
+import {calculateTripTime} from '../utils.js';
 
-const HOURS_COUNT = 24;
-const MINUTES_COUNT = 60;
-const DAY_MILLISECONDS_COUNT = 86400000;
-const HOUR_MILLISECONDS_COUNT = 3600000;
-const MIN_MILLISECONDS_COUNT = 60000;
-
-export const createTripEventTemplate = (point) => {
+export const createTripEventTemplate = (point, currentDay) => {
 
   const pointType = point.pointType;
   const pointDestination = point.destination;
@@ -15,38 +10,18 @@ export const createTripEventTemplate = (point) => {
   const minutesDeparture = point.minutesDeparture;
   const hoursArrival = point.hoursArrival;
   const minutesArrival = point.minutesArrival;
+  const tripPrice = point.price;
+
+  const year = currentDay.year;
+  const month = currentDay.month;
+  const day = currentDay.day;
 
   const generateRandomTime = (start, finish) => {
     return `${correctDateFormat(start)}:${correctDateFormat(finish)}`;
   };
 
-  let calculateTripTime = (departure, arrival) => {
-    let firstDate = departure.toString();
-    let secondDate = arrival.toString();
-
-    let getDate = (string) => new Date(0, 0, 0, string.split(`:`)[0], string.split(`:`)[1]);
-    let different = (getDate(secondDate) - getDate(firstDate));
-    let differentRes; let hours; let minuts;
-    if (different > 0) {
-      differentRes = different;
-      hours = Math.floor((differentRes % DAY_MILLISECONDS_COUNT) / HOUR_MILLISECONDS_COUNT);
-      minuts = Math.round(((differentRes % DAY_MILLISECONDS_COUNT) % HOUR_MILLISECONDS_COUNT) / MIN_MILLISECONDS_COUNT);
-    } else {
-      differentRes = Math.abs((getDate(firstDate) - getDate(secondDate)));
-      hours = Math.floor(HOURS_COUNT - (differentRes % DAY_MILLISECONDS_COUNT) / HOUR_MILLISECONDS_COUNT);
-      minuts = Math.round(MINUTES_COUNT - ((differentRes % DAY_MILLISECONDS_COUNT) % HOUR_MILLISECONDS_COUNT) / MIN_MILLISECONDS_COUNT);
-    }
-
-    if (hours <= 0) {
-      return `${correctDateFormat(minuts)}М`;
-    }
-
-    return `${correctDateFormat(hours)}H ${correctDateFormat(minuts)}М`;
-  };
-
   const departure = generateRandomTime(hoursDeparture, minutesDeparture);
   const arrival = generateRandomTime(hoursArrival, minutesArrival);
-
   const tripTime = calculateTripTime(departure, arrival);
 
   return (
@@ -59,15 +34,15 @@ export const createTripEventTemplate = (point) => {
 
         <div class="event__schedule">
           <p class="event__time">
-            <time class="event__start-time" datetime="2019-03-18T${departure}">${departure}</time>
+            <time class="event__start-time" datetime="20${year}-${month}-${year}T${day}">${departure}</time>
             &mdash;
-            <time class="event__end-time" datetime="2019-03-18T${arrival}">${arrival}</time>
+            <time class="event__end-time" datetime="20${year}-${month}-${day}T${arrival}">${arrival}</time>
           </p>
           <p class="event__duration">${tripTime}</p>
         </div>
 
         <p class="event__price">
-          &euro;&nbsp;<span class="event__price-value">20</span>
+          &euro;&nbsp;<span class="event__price-value">${tripPrice}</span>
         </p>
 
         <h4 class="visually-hidden">Offers:</h4>
